Finished section Authorization

+++ src/features/ticket/actions/delete-ticket.ts
@@ -3,12 +3,30 @@
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { setCookieByKey } from "@/actions/cookies";
-import { fromErrorToActionState } from "@/components/form/utils/to-action-state";
+import {
+  fromErrorToActionState,
+  toActionState,
+} from "@/components/form/utils/to-action-state";
+import { getAuth } from "@/features/auth/actions/get-auth";
+import { isOwner } from "@/features/auth/utils/is-owner";
 import { prisma } from "@/lib/prisma";
-import { ticketsPath } from "@/path";
+import { signInPath, ticketsPath } from "@/path";
 
 export async function deleteTicket(id: string) {
+  const { user } = await getAuth();
+
+  if (!user) {
+    redirect(signInPath());
+  }
+
   try {
+    if (id) {
+      const ticket = await prisma.ticket.findUnique({ where: { id } });
+      if (!ticket || !isOwner(user, ticket)) {
+        return toActionState("ERROR", "Not Authorized");
+      }
+    }
+
     await prisma.ticket.delete({ where: { id } });
   } catch (error: unknown) {
     return fromErrorToActionState(error);
